swapchain: add resizing API
This is used both for updating the size and querying the size. I don't
want to make these separate functions because it should be painfully
obvious that the size you get may not be the size you request.

This allows libplacebo to work on wayland, which mediates the concept of
swapchain resizing to protocols like xdg_shell that mesa/vulkan can't
know anything about (by design).