vulkan: permit non-coherent memory
Since we control to a very fine degree when and where memory flushing
and invalidation happens anyway, there's no reason for us to require
coherent memory.

This is a very slight performance gain on my tested platforms, but more
importantly, this allows libplacebo to work on platforms which only have
noncoherent memory (such as some iGPUs).