shaders/av1: release under MIT
dav1d people want this to be permissive so they can adapt it for
documentation purposes. I have no problems releasing this specific file
under a permissive license.

Note that since both the header and source code heavily depend on the
LGPLv2.1+ parts of libplacebo, this licensing change isn't terribly
useful unless you're just interested in copying fragments of the glsl
code or some of the C helper functions.