config: replace PL_FIX_VER/PL_VERSION by runtime functions
Since the fix version is not part of the ABI, we can't use a header
define for it. (And indeed, users can't use this information as part of
any preprocessor check - nor should they need to)

I won't bother bumping the API version for this change because it's
unlikely to result in any breakage.