renderer: automatically adjust deband grain scale
This fixes blowing up with excessive grain on e.g. HDR sources. I
decided not to change the signature of the deband shader itself at this
time, since the deband shader doesn't really know anything about the
meaning / interpretation of what it's debanding. This is the user's
responsibility when using the shader directly.

Fixes #59.