gpu: turn undefined behavior into explicit errors
Wherever possible/appropriate. Bumped the API version for this because
it is quite a significant change in semantics. At last, API users are
freed from the responsibility of needing to duplicate every check under
the sun...

Annoyingly, some functions return `void`, so there's not much we can do
except print an error and hope the user notices it before it's too late.
Maybe we should introduce some sort of internal `error bit` allowing the
user to later check for presence of an error?