vulkan: differentiate between buffer and texture representations
The only emulation code path is the buffer<->texture transfer, so we
need to create vertex/texel buffers with the format being emulated, not
the format being used for textures. This might change in the future, if
we create an emulation path for pl_buf_write/read (but note that
host-mapped buffers would be disallowed!).

For now, simply exclude all buffer-related caps if there isn't an
explicit buffer format specified for emulated formats, since this allows
us to easily exclude texel caps for "nonsensical" formats like r16f
(r32f would be the correct format to use for this type of buffer).

Also add some extra validation around texel size assumptions.