gpu: fix stride/alignment of std430 vectors/matrices
Yet another bug in this accursed set of functions. The base alignment of
std430's 3-component vectors is still 4*unit, even though the spec says
that this is not rounded up to vec4 - in this case, the vec4 alignment
is already implicit in the base alignment of a vec3.

Additionally, matrices/vectors always have their stride rounded up to
the base alignment, which in the case of e.g. mat3 specifically, is
4*float, not 3*float - the only difference to std140 is that it's not
explicitly also rounded up to a multiple of vec4.

Also add some more assertions just because, and fix a wrong test case.