dispatch: place variables inside PCs more greedily where possible
Instead of spilling all matrices and static vars over into UBOs, at
least *try* and place them into PCs as well.

Note: This has the minor quirk of generating PC blocks that can be
out-of-order. It's not a huge deal, since we explicitly annotate an
offset everywhere, but it's at least worth pointing out.

Closes #73.