vulkan: actually rotate queues only once per frame
Either on pl_gpu_flush(), or on pl_swapchain_submit_frame(). That should
hopefully about cover all the common cases of when we actually want to
rotate queues.

In the past, it was possible that this could happen at chaotic times -
such as when exporting shared buffers - thus making using more than one
queue actually be a performance loss.

Also add a bit of debugging for accountability.