vulkan: expose pl_vulkan_choose_device
This is a useful helper function for scripts. I debated whether or not
to just copy/paste the options struct or if we should retroactively
change the existing one for aesthetics, but I'm tired of breaking the
API for an insufficiently good reason. Plus, we need to set a different
instance anyway, so this isn't too bad.

Another alternative would have been to re-use the actual
`pl_vulkan_params` but this might be even more confusing. *Shrug*

Closes #74.