vulkan: expose PCI bus address for interop
The great thing about standards is that there are so many to choose
from.

While nvidia uses UUIDs to match vulkan and cuda devices, mesa uses
PCI bus addresses to match with DRM devices - primarily because DRM
only provides the PCI bus address as identity information. So, if
we want to ensure that a client can match the vulkan and DRM devices,
we need to expose the PCI information.