shaders/colorspace: actually implement pl_color_adjustment.gamma
This was completely forgotten about, which might have to do with the
fact that it plain simply doesn't really belong in `pl_color_adjustment`
to begin with. All of the other parameters go into linear YUV->RGB
conversion, but gamma is nonlinear and ideally needs to be done after
linearization. Then again, so do all of the other steps. The status quo
of this entire struct is that we just treat it as non-linear perceptual
color adjustments, so just apply the gamma immediately after decoding
color.

I have no idea how horribly this breaks for images with an alpha
channel. I guess *somebody* will find out.

Closes https://github.com/haasn/libplacebo/issues/68