shaders/av1: refactor and split up `pl_grain_params`
Move the raw data (from the header itself) into a separate struct
`pl_av1_grain_data` and, in libplacebo style, have the `params` include
the fields that currently spill over into the parameter list.

The split is intended to enable attaching `pl_av1_grain_data` as
metadata to the renderer's `pl_image` struct without needing to
introduce reduncancy with `repr`, `channels` etc.

This also changes the signature (adds a return value) and introduces a
new helper function for testing whether or not grain needs to be applied
at all.