opengl: actually implements params.debug
This was completely forgotten about. Also document the behavior.

Tremendous care needs to be taken here because OpenGL provides no way to
destroy debug report callbacks after they're no longer needed, which due
to the library nature of libplacebo means that this function can be hit
well after the destruction of the `pl_opengl`, or even the `pl_context`
it was initialized from.

We pay very close attentino to the lifetime of the `pl_context` and its
association with the debug report callback.