talloc: unlock mutex before destroying it
This suppresses a validation issue in tsan which complains about
destroying locked mutexes for some reason, even though the pthread man
page claims this is fine.

Apparently it causes issues if other threads are still waiting on this
mutex, but if we're destroying a mutex that, to me, implies we already
know nobody is waiting on it - otherwise you'd have the exact same race
condition in between unlocking it and destroying it.

It makes absolutely no sense, but whatever makes tsan happy, I guess.