opengl: remove debug callback pthread madness
I was under the mistaken impression that calling glDebugMessageCallback
twice would add two callbacks (like in vulkan land), rather than
overriding the callback to the second value. But OpenGL's callback
mechanism is based on mutable global state (of course...).

So we can safely just un-set the callback when we're done with it. And
given that OpenGL is so globally mutable in general, I'm not sure it
even remotely makes sense for us to be worrying about refcounting and
multiple `pl_opengl` instances. Just nuke it all and make it clear that
this is undefined behavior.