vulkan: implement VK_EXT_hdr_metadata
This actually exists on recent versions of AMDVLK, and is also pretty
widespread on Windows. Note that we only support this as advertised if
the swapchain is actually configured to a HDR mode, because anything
else is a bit pointless.

We also cache the HDR metadata and re-apply it after swapchain
recreation, so that this state is persistent from the point of view of
the API user.