shaders/colorspace: improve pl_shader_encode_color skipping
The current code doesn't skip the cmat even if the only operation that
needs to be applied is the independent alpha handling.

Instead of only returning if the entire function is a no-op, emit the
parts of the code that can always be safely emitted and only skip the
cmat if it's a no-op.