vulkan: enable loading higher vulkan versions
This doesn't really do anything, and we haven't bumped up our minimum
version requirements, but provides the necessary public-facing API
changes needed to make use of this functionality in the future.

Users may restrict the API version, although the usefulness of doing
this on a device level is questionable at best since we enable
extensions on an as-available basis anyway. (More useful would be an
extension blacklist!)