renderer: refactor plane state representation
This will make it easier to enable support for additional preprocessing
steps, including custom user shaders, moving forwards.

This refactor completely eliminates `image->src_rect` from the signal
path (after the initial plane rects are determined), with the only major
change in logic being that planes are no longer aligned relative to the
ref plane, instead all planes are aligned directly relative to the
target rect. In theory this could introduce some loss of information
near misaligned luma edges, but since misaligned luma is just a
hypothetical case anyway (permitted by our pl_plane abstraction) it
shouldn't cause any differences in practice.

In theory we could get away with enforcing shift=0 for the refplane, but
why bother if we can also just support that edge case without too much
difficulty?

This commit also deprecates the `width/height` fields from pl_image.
These were used in the past to determine the refplane, but since we now
determine this based on the plane type (and failing that, order), it's
irrelevant. I could just remove the fields but because this is just a
removal of a relatively commonly used struct member, I decided to
instead just mark it as deprecated.