filters: add proper nearest neighbour filter
The built-in 'box' filter doesn't correspond to true nearest neighbour
sampling because it's resizeable and thus breaks when downscaling. To
correct this, introduce a new copy (called 'nearest') which is
functionally identical but marked as non-resizeable, and make the filter
generation code ignore filter scale for non-resizeable filters.

Test it to make sure it actually doesn't blend any pixels at any scale.