vulkan: allow specifying device UUIDs for device selection
We already support UUIDs for interop purposes, it doesn't take much to
also allow them for device selection purposes. This makes obvious sense
since we use UUIDs for interop purposes, so in the case that the user
already has e.g. an external hwdec API telling them what UUID to use,
they also need to pair that with the correct vulkan device on the
libplacebo side of things.

Notably, this commit extends the VK_KHR_get_physical_device_properties2
extension requirement to pl_vulkan_choose_device as well, but that's not
a huge departure from the status quo because pl_vulkan_create already
requires it. We just need to document it as such.

Add some extra printfs of the UUIDs here and there to make this stuff
clearer.