filters: compute rows by real radius, not rounded row size
This prevents excessive stretching/blurring/shifting of the kernels in
cases where the row size is not an exact multiple of the radius (such as
is the case when performing anti-aliased downsampling).

Tested by making sure the output matches imagemagick. (Can we get proper
test cases for this kind of stuff already?...)

Fixes #72