renderer: ensure image->src_rect is valid
Since we can't infer this right away, do it first thing inside
pass_read_image. This simplifies the code a bit and also the logic, by
making sure `image->src_rect` actually means what it's supposed to mean.

As an aside, also fix a missing initialization of `cur_img.rect` after
scaling is complete, which is kinda pointless in the current code but
technically the correct thing to do.