filters: fix antialiased downsampling.. again
The new code was still wrong for some downscaling ratios. The bottom
line is that it's correct to use the rounded value here, but the actual
issue was that it was being rounded to an odd number of samples rather
than a multiple of 2, which broke the centering alignment assumptions
made by both the code and the caller.

Fixes https://github.com/haasn/libplacebo/issues/72 again and hopefully
for the last time.

This is probably not the super-duper theoretically optimal solution to
the issue, because it means we may require one more sample than would
strictly speaking be sufficient, but whatever. The alternate is
hard-coding an offset of 1 for odd row sizes, I guess, but that feels
kinda sketchy. I'd rather make the code match what mpv and IM are doing
internally.