Revert "filters: add proper nearest neighbour filter"
This reverts commit b0e6d362b16b2c6930479b7bda818d8c5d4883f3.

This commit made the oversight of not all downsampling kernels being
resizeable, and in particular, it being resizeable or not has nothing to
do with our ability to stretch the filter scale. This commit effectively
broke anti-aliased downsampling for e.g. bcspline-family filters, such
as the default mitchell filter, which is fairly significant.

I think the only proper way to get nearest neighbour downsampling is to
set box as the kernel *and* also disable anti-aliasing.