vulkan: drop the use of shaderStorageImageExtendedFormats
I have no idea why we were enabling this. I'm pretty sure it's because
some validation layer in the past told me it was required. But the spec
itself disagrees, saying instead that enabling this has effectively no
effect because the relevant formats will already advertise storage
support when queried.