config: undefine unset symbols
This is more convenient since it allows users to treat "libplacebo too
old to support feature" and "libplacebo compiled without feature" as the
same thing, by simply testing for #ifdef.

Rather unfortunately, though, this is not quite backwards compatible -
users will need to continue using #if for a while, which also means that
they will get warnings when testing for unset symbols since the
preprocessor will default them to 0.

But moving forwards, this is the design that should probably make users
happiest in the long run.