dispatch: add support for pl_timer queries
Rather than add yet another field to the growing signature of these
functions, I decided to do the sensible thing and just params-ify it.
(When will I learn my lesson and just use params structs to begin with?)

While this rather annoyingly requires completely breaking the API of a
relatively common function, we're fortunately saved by the fact that I
expect the majority of pl_dispatch uses to be libplacebo-internal.

This commit also adds the use of such timers to the benchmark, giving us
more accurate reporting of the raw GPU execution time.