renderer: minor struct img refactor
Rather than tying struct img to the specifics of pass->cur_img, I wanted
to loosen this up a bit and re-use it for the plane state. To this end,
I parametrized one key function, `pass_hook`, by the actual `img` to
use, and got rid of some redundant fields.

This commit doesn't (or shouldn't) change the behavior yet, it's mostly
to disentangle the behavior-changing bits of this refactor from the
merely refactoring ones.