shaders/custom: don't force OUTPUT fbo for szexprs
09834c521 introduced a regression here that caused an unnecessary FBO
indirection to occur for the (relatively common) case of user shaders
consulting the dimensions of OUTPUT in their RPN expressions.

OUTPUT apparently behaves weirdly. mpv should probably document this
better. Even better would be to have had different names for the OUTPUT
hook stage and the OUTPUT magic constant for RPN exprs. But it's
probably too late for that now. Unless we want to rename the OUTPUT
stage, which we totally could do.