renderer: simplify offset rounding / plane alignment
The current logic could end rounding the plane up in both directions,
inadvertently introducing an upscale by 1 pixel to the refplane, which
not only forced an FBO indirection but also lowered the quality due to
the effective resample. Furthermore, the code for adjusting the rect by
the `rc` was wrong, since it failed to scale the "affine" part of the
transformation down to the coordinate space of the plane.

Simplify and fix this by only rounding the offset (making sure to always
round towards 0 to have the best chance of "doing the right thing"), and
also correctly scaling this offset when calculating the offsets for the
individual planes.