vulkan: generate boilerplate instead of hand-writing it
Turns out I need to generate even more boilerplate than just this, so
just write a shitty python script (inspired by RADV/mesa) to do the job.

Makes python part of the build dependency of libplacebo, but meson
already depends on python so nothing has changed.

The CI URL for aarch64 needs to be updated to pull in python3-mako.