vulkan: clean up VK_KHR_swapchain handling
This extension was treated as global in the past, but now that we have
a proper extension framework we should just handle it like any other
extension. Solves a very concrete issue where dependencies on e.g.
VK_EXT_hdr_metadata were not satisfied if the user did not happen to
enable this extension.

Also check if the extension is loaded when we attempt actually creating
a swapchain.