renderer: refactor pl_render_target.dst_rect
This is now pl_rect2df instead of pl_rect2d, to make it easier to use
the pl_rect2df_aspect* series of functions, especially without requiring
the hacky rounding integer versions of them. Delete those and add some
needed helper functions as well.

Rewrite the fix_rects code to crop `src_rect` for any fractional
offset in the `dst_rect`, and also for regions of the `dst_rect` that
lie outside the target fbo.

Also fix a bug in the img->w/h calculation for cropped planes.