renderer: fix plane pass dimensions
Turns out the correct handling img.w/h isn't as simple as setting it to
either "the plane size" or "the src_rect size". We have to set it to the
plane size for the plane shaders, but the src_rect size for the plane
scalers. How annoying. Maybe this field needs to be reworked/figured out
somehow.