dispatch: attach the appropriate memory qualifiers
We consider all descriptors to be 'restrict'-qualified, because there
just plain is no use case we have for aliasing different descriptors.
Readonly/writeonly is alreayd handled by the descriptor access mode, so
no need to duplicate those memory qualifiers either.

The only ones left are 'coherent', which is required for the peak
detection shader's memory barriers to work, and 'volatile', which we
don't use. Might have a use for it in the future though, especially for
development purposes.

Slight performance gain for compute shaders (especially AV1 grain), but
more importantly, fixes a bug.