renderer: deprecate/remove pl_image.signature
I decided the current semantics of this struct don't really make sense.
There's no point in forcing _every_ user to constantly be setting unique
signatures. Most users don't care, and in particular, I've made the same
mistake in my own code more often than not.

Furthermore, it's incredibly disingenuous to imply that libplacebo
caches images to speed up redraws when it effectively does not. I'm not
even sure if I think this is a good thing. It's probably a better idea
for the user to do this, because they're the only ones that have a clear
understanding of which frames will be redraws and which will not.

The bottom line is, I've decided that pl_render_image does not and
should not need to cache. So remove all references to the frame cache.