vulkan: refcount pl_tex
See https://code.videolan.org/videolan/libplacebo/-/issues/87 for a
discussion of why.

Upon re-consideration I've determined that I'm not comfortable exposing
`pl_tex_poll` because I still fail to see a sane use case for this. The
only reasonable scenario is when using it together with `pl_tex_export`,
but that function already requires synchronization using semaphores.

I'd have to change the signature to make the `sync` optional, but I'm
not comfortable doing this without a good motivating example, because I
think omitting the `sync` and using CPU-based waits is a last resort.
(And that said, it would have to be done in both directions or risk UB)