gpu: dispatch: vulkan: allow formatless storage images
Using the ReadWithoutFormat/WriteWithoutFormat device features. This
suppresses the annoying "Storable format has no matching format
qualifier" warning spam, and makes it a proper warning.

Technically an API break but I decided it's not major enough to bump the
API version, because it's sort of a strict upgrade, and I don't think
this API has any users besides mpv (which handles it correctly already).