dispatch: allow dispatching compute shaders with outputs
I realized that for the purposes of e.g. `pl_shader_detect_peak`,
dispatching it as a compute shader while ignoring its output can be
useful. Based on this shader alone, there's no reason to force compute
shaders to have no output.

Not bumping the API version for this because it doesn't represent any
change in the signature, and only enables extra semantics that are
currently useless anyway. (The next commit will fix that)