spirv/glslang: fix build on 8.13.3743
The API change referenced in 5230568 was introduced in between patch
level 3743 and patch level 3763. The current check breaks the tagged
release 8.13.3743 which does not include the change yet.

Bump the check up to 3763 to get around this. In classic glslang style,
this means that there are a few commits which will always fail to
compile no matter what we do... but that's the reality we have to live
with when building against glslang. It's either that or they figure out
how to bump their patch level on breaking API changes.