shaders: generalize sh_lut to support other variable types
The AV1 code in particular really wants to be generating integer LUTs,
and there's no clear reason why `sh_lut` shouldn't support that use
case. It's a relatively easy thing to support.

We move the format probing code to happen before method selection, so
that we can automatically rule out textures if there's no matching
texture format for the desired configuration (e.g. for odd-sized
formats, or integer formats on older versions of OpenGL).

We also make sure that integer textures are created with at least 32 bit
depth, because the precision loss is absolutely not acceptable for
integer LUT data.