vulkan: pick fewer swapchain usage flags
Enabling everything under the sun really goes against the spirit of
these flags existing to begin with, and in particular, introduces
unnecessary image layout transitions because the pl_pass code doesn't
know wtf to do.

Instead, just enable the ones we actually care about, which ends up
causing the layout transitions to be a no-op. (In theory it would be
cool if we could somehow figure out if we're actually rendering to a
swapchain image, but oh well)