gpu: add support for generic vertex buffers
Because of annoying API dependency issues (vertex buffers depend on
pl_buf which depends on pl_buf_read/pl_tex_upload, which depends on PBO
support), we can't use pl_pass_run_vbo for OpenGL, so instead we have to
manually create and update the vertex buffer. (Not a huge deal since
this is what the code did previously)

It also means we miss out on vertex buffer reuse for OpenGL, but oh
well. I guess the driver could already be doing this internally.