shaders/colorspace: read detected peak directly from ssbo
With the recent series of refactors to the vulkan malloc layer,
host-visible device-local memory types exist and are allocatable, so we
can directly serve host-readable uniform buffers.

For the scenarios in which it's not possible, working around it should
probably be done inside the pl_gpu, not the application code. (i.e.
'host visibility emulation')