gpu: add preliminary API support for DRM format modifiers
This is still a pretty bad hack-patch as of currently, because no driver
actually implements the drm format modifier extension. But this way of
doing it at least allows us to differentiate between linear and
non-linear, which we assume (blindly) is equal to optimal, and is needed
to get vaapi hwdec working on AMD.

We also get rid of the plane offset check because this also conflicts
with the requirements of drm format modifiers, which we again can't
respect properly. We already suppress validation errors for the image
bind, and it works in practice.