vulkan/malloc: round imported pointers to page boundaries
This allows us to bypass the page-alignment restriction on host pointer
imports, by simply sufficiently extending the host pointer base, the
buffer offset, and the memory size in the respective direction. Thus
ensuring that our memory import is always page-aligned.

This *should* technically be safe, because the MMU can only enforce
virtual memory access safety on a per-page granularity, and our code
should never end up reading outside the bounds of a vk_memslice. But on
the other hand, what we're doing is absolutely insane. Beware nasal
demons. I only wrote this logic because I enjoy sharing an address space
with a malevolent agent of chaos.

As an aside, also fix some errors related to imported buffer size
calculation and alignment validation that I noticed along the way.