gpu: use host-imported pointers for pl_tex_download_pbo
This allows such tex transfers to avoid an extra memcpy in most cases,
except where the pointer happens to be horrifically misaligned with
respect to the texel size - but in these cases, the alignment-fixing
memcpy will happen inside VRAM (PL_BUF_MEM_DEVICE), which should still
be faster than doing an extra memcpy in RAM.

Also, I realized it makes no sense to have tex_download_pbo use a buffer
pool at all, because it's synchronous anyway - there can only ever be
one buffer. And doing it this way avoids code duplication between the
import branch and the non-import branch.

Side note: We could do the same for pl_tex_upload_pbo with the same
justification, but I decided to test the waters with this commit first.