opengl: implement exporting and importing through EGL
Implements PL_HANDLE_DMA_BUF for opengl.

This commit also introduces a public-facing concept of DRM fourccs,
which are useful/needed by OpenGL to mediate the texture format. We
blend this into the pl_fmt concept because of parallels to how this is
done in other APIs, and because `pl_fmt` is already required to specify
upon import.

See !116 for discussion.

Co-authored-by: Niklas Haas <git@haasn.xyz>