shaders: refactor away parse_user_shader
This function was annoying and served no real purpose, because it lives
in the same file as the only place that needs it. It's a left-over
design quirk from mpv.

In addition to deleting this function, the custom_shader_tex has been
replaced by the more generic pl_shader_desc, which allows us to create
and attach more resource types than just textures, via the same
mechanism.

This commit is non-functional, it's a pure refactor.