shaders: add rudimentary support for storage images
This allows shaders to have persistent state across images. We
piggy-back the storage texture support off the existing texture support,
by simply adding a new directive //!STORAGE which marks the image as
a storage image. (It's still bound using the existing //!BIND, which now
simply causes the resulting texture to get bound as an image2D instead)

This still has the limitation that the texture size needs to be static,
whereas ideally it would be cool if we could specify the storage size as
an RPN expression, but it's an easy way to implement what we want. (And
in practice, you can just make the texture larger than it needs to be)

Partially implements https://code.videolan.org/videolan/libplacebo/-/issues/110