shaders: add support for custom BUFFER resources
This allows the use of uniform buffers and storage buffers,
respectively. The interface is similar to //!TEXTURE. Currently, the
only supported directive is //!STORAGE, which marks it as a readwrite
storage buffer, and //!VAR <type> <name> which declares a variable
(similar to writing `type name;` in GLSL).

Also add a rudimentary test case, that doesn't actually test anything.

Closes https://code.videolan.org/videolan/libplacebo/-/issues/110