utils/libav: add helpers for mangling avframes into libplacebo
These structs and enums are designed in not-quite-compatible ways, and
the mapping is not at all intuitive or obvious. This should hopefully
make life a billion times easier for people wanting to use libplacebo
together with FFmpeg/libav* (i.e., basically the default use case).

Also includes an ultra-high-level wrapper that just takes an AVFrame *
and straight up uploads it to the GPU, in `pl_image` format, ready to be
consumed by `pl_renderer`.

This entire set of helpers is done as a single-header library to avoid
depending on libavutil, which would be a *really* terrible idea since
libavfilter also wants to depend on libplacebo. This single-header
nonsense is the easiest way to avoid a cyclic dependency. I promise I
won't make it a habit. Besides, it's annoying having to do things this
way, because it means I can't re-use any of the fancy logging functions
from pl_ctx. Oh well. I hope the user knows what they're doing.